BoundedInterval: rearrange code + fix duration
