+++ modules/tests/shared/src/test/scala/lucuma/core/math/BoundedIntervalSuite.scala
@@ -8,13 +8,20 @@ import lucuma.core.arb.ArbTime
 import lucuma.core.math.BoundedInterval
 import lucuma.core.math.arb.ArbInterval
 import lucuma.core.optics.Spire
+import lucuma.core.optics.laws.discipline.FormatTests
+import lucuma.core.optics.laws.discipline.SplitEpiTests
+import lucuma.core.syntax.time.*
+import lucuma.core.tests.ScalaCheckFlaky
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalacheck.Gen
 import org.scalacheck.Prop.*
 import org.typelevel.cats.time.*
 import spire.math.Interval
 
+import java.time.Duration
 import java.time.Instant
+import java.time.LocalTime
+import java.time.ZoneId
 
 class BoundedIntervalSuite  extends munit.DisciplineSuite with IntervalGens {
   import ArbInterval.given
@@ -93,7 +100,7 @@ class BoundedIntervalSuite  extends munit.DisciplineSuite with IntervalGens {
       forAll(
         distinctZip(instantUntilEndOfInterval(i), instantFromStartOfInterval(i))
       ) { instants =>
-        Spire.openUpperIntervalFromTuple[Instant].getOption(instants).foreach { other =>
+        BoundedInterval.openUpperFromTuple[Instant].getOption(instants).foreach { other =>
           val join = i.join(other)
           assertEquals(join.map(_.lower), List(i.lower, other.lower).min.some)
           assertEquals(join.map(_.upper), List(i.upper, other.upper).max.some)
@@ -116,10 +123,35 @@ class BoundedIntervalSuite  extends munit.DisciplineSuite with IntervalGens {
       ) { instantsOpt =>
         assert(
           instantsOpt
-            .flatMap(Spire.openUpperIntervalFromTuple[Instant].getOption)
+            .flatMap(BoundedInterval.openUpperFromTuple[Instant].getOption)
             .forall(other => i.join(other).isEmpty)
         )
       }
     }
   }
+
+  test("ToFullDays".tag(ScalaCheckFlaky)) {
+    forAll { (i: BoundedInterval[Instant], z: ZoneId, t: LocalTime) =>
+      val allDay = i.toFullDays(z, t)
+      assert(i.isSubsetOf(allDay))
+      // We can't comparte LocalTimes directly since the LocalTime may not exist at
+      // the given day if there was a DST transition.
+      val start  = allDay.lower.atZone(z)
+      assertEquals(start, start.`with`(t))
+      val end    = allDay.upper.atZone(z)
+      assertEquals(end, end.`with`(t))
+      assert((allDay -- i).forall(_.asInstanceOf[BoundedInterval[Instant]].duration < Duration.ofDays(1)))
+    }
+  }
+
+  // Optics
+  checkAll(
+    "BoundedInterval.openUpperFromTuple",
+    FormatTests(BoundedInterval.openUpperFromTuple[Int]).format
+  )
+
+  checkAll(
+    "BoundedInterval.closedFromTuple",
+    FormatTests(BoundedInterval.closedFromTuple[Int]).format
+  )
 }
